Removed mul_mod in pow_mod.

+++ 007.cpp
@@ -4,7 +4,7 @@
 
 int main()
 {
-    int prime = 1;
+    unsigned long long prime = 1;
     for (int i = 0; i < 10001; i++)
         prime = util::get_next_prime(prime);
